:wrench: [refactor] (#88) data class 정리 및 이전의 API 인터페이스 제거
